[Fix] Change data type for matrix entries to accept float values

+++ lib/matrix/matrix.h
@@ -2,7 +2,7 @@
 #define MATRIX_H_
 
 typedef struct {
-    int **entries;
+    double **entries;
     int row;
     int col;
 } Matrix;
